+++ src/main/java/edu/hm/hafner/analysis/LookaheadParser.java
@@ -25,6 +25,8 @@ public abstract class LookaheadParser extends IssueParser {
     private static final String ENTERING_DIRECTORY = "Entering directory";
     private static final Pattern MAKE_PATH
             = Pattern.compile(".*make(?:\\[\\d+])?: " + ENTERING_DIRECTORY + " [`'](?<dir>.*)['`]");
+    private static final String NINJA_PREFIX = "-- Build files have";
+    private static final Pattern NINJA_PATH = Pattern.compile(NINJA_PREFIX + " been written to: (?<dir>.*)");
 
     private final Pattern pattern;
 
@@ -52,6 +54,12 @@ public abstract class LookaheadParser extends IssueParser {
                     Matcher makeLineMatcher = MAKE_PATH.matcher(line);
                     if (makeLineMatcher.matches()) {
                         builder.setDirectory(makeLineMatcher.group("dir"));
+                    }
+                }
+                else if (line.contains(NINJA_PREFIX)) {
+                    Matcher ninja = NINJA_PATH.matcher(line);
+                    if (ninja.matches()) {
+                        builder.setDirectory(ninja.group("dir"));
                     }
                 }
                 else if (isLineInteresting(line)) {